Clean up

+++ src/mr/view/screen/ScreenRenderer.java
@@ -25,7 +25,7 @@ public class ScreenRenderer {
 				int x = i%GameConstant.WIDTH;
 				int y = (i-x)/GameConstant.WIDTH;
 				
-				Coordinate coord = new Coordinate((float)x*GameConstant.TILE_SIZE, (float)y*GameConstant.TILE_SIZE);
+				Coordinate coord = new Coordinate(x*GameConstant.TILE_SIZE, y*GameConstant.TILE_SIZE);
 				// -1 to handle 0 as empty tile (therefore index 1 in the level = 0 in the array)
 				images.add(new RenderingImage(coord, level.getTiles().get(tiles[i]-1)));
 			}
